requirements.txt
# Runtime is intentionally zero-dependency. The integration layer only relies
# on the Python standard library so the package stays trivially embeddable.
#
# Optional extras pulled in by individual backends:
#   google-genai   # only needed for the Gemini cloud fallback (filter_input/output and StubBackend work without it)
#
# Cactus + functiongemma weights are installed via the cactus CLI:
#   cd ../cactus && source ./setup
#   cactus build --python
#   cactus download google/functiongemma-270m-it --reconvert
google-genai>=1.0
